requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
dateparser==1.2.1
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.3.0
psycopg2-binary==2.9.10
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
python-binance==1.0.28
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.3.0
websockets==15.0
yarl==1.18.3
